az-functions/analis-imagens-blob/requirements.txt
anyio==4.2.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.29.6
azure-functions==1.18.0
azure-storage-queue==12.9.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
exceptiongroup==1.2.0
idna==3.6
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
pycparser==2.21
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
typing_extensions==4.9.0
urllib3==2.1.0
